telegram_bot/requirements.txt
python-telegram-bot[job-queue]>=21.0
google-generativeai
httpx
redis
python-dotenv
pydantic
apscheduler
